[FLINK-13877][hive] Support Hive version 2.1.0 and 2.1.1
Support Hive 2.1.x versions (2.1.0 and 2.1.1).

This closes #9547.
