+++ flink-connectors/flink-connector-hive/src/main/java/org/apache/flink/table/catalog/hive/client/HiveShimLoader.java
@@ -35,6 +35,8 @@ public class HiveShimLoader {
 	public static final String HIVE_VERSION_V1_2_0 = "1.2.0";
 	public static final String HIVE_VERSION_V1_2_1 = "1.2.1";
 	public static final String HIVE_VERSION_V1_2_2 = "1.2.2";
+	public static final String HIVE_VERSION_V2_1_0 = "2.1.0";
+	public static final String HIVE_VERSION_V2_1_1 = "2.1.1";
 	public static final String HIVE_VERSION_V2_3_0 = "2.3.0";
 	public static final String HIVE_VERSION_V2_3_1 = "2.3.1";
 	public static final String HIVE_VERSION_V2_3_2 = "2.3.2";
@@ -59,6 +61,12 @@ public class HiveShimLoader {
 			}
 			if (v.startsWith(HIVE_VERSION_V1_2_2)) {
 				return new HiveShimV122();
+			}
+			if (v.startsWith(HIVE_VERSION_V2_1_0)) {
+				return new HiveShimV210();
+			}
+			if (v.startsWith(HIVE_VERSION_V2_1_1)) {
+				return new HiveShimV211();
 			}
 			if (v.startsWith(HIVE_VERSION_V2_3_0)) {
 				return new HiveShimV230();
